refactor(label_placement): extract tests to external file; use Alignment + rename truncate helper
- Move inline mod tests to label_placement_tests.rs (207 → 207 LOC,
  previously 498 — resolves File Size Lint CI failure on PR #578).
- Replace private Side enum with ratatui::layout::Alignment.
- Rename truncate_with_ellipsis → truncate_label_to_cells to avoid
  shadowing src/util/formatting.rs::truncate_with_ellipsis which uses
  bytes + "..." semantics vs chars + "…".
